\name{EpipwR.data}
\alias{EpipwR.data}
\title{EpipwR.data: Reference data for EpipwR packages}
\description{
This package provides reference data required for the EpipwR package.
Additional information about the source of each reference
data set can be found in the vignette: vignette("EpipwR.data")
}
\value{EpipwR.data gives the user access to experiment data
needed for the EpipwR R package}
